// app/pages/LeaveRequestSection/LeaveRequestSection.interface.ts
export interface ILeaveRequestForm {
	fio: string;
	companyName: string;
	city: string;
	email: string;
	phone: string;
	objectType: string;
	qualityType: string;
	distanceCameras: string;
	roomSize: string;
	archiveSize: string;
	distanceSupply: string;

	purpose: string;

	numberCameras: string;

	comment: string;
}